requirements.txt
streamlit>=1.31.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
bcrypt>=4.0.1
pandas>=2.0.0
altair>=5.0.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2
requests>=2.31.0
pydeck>=0.9.0
python-dotenv>=1.0.0
